Add main UpdateBaseElementHrefOnly option and change the Rewriters option
The Rewriters option no longer replaces the built-in rewriters.  Now it adds your
custom rewriters as the first rewriters, but still falls back to the built-in ones
if a file is not handled by a default rewriter.

Also add more unit tests that includes at least one test for each option.